(chore) refactor all state updates to go via pubsub
display now listens on its own pubsub channel for state updates, rather
than being on the same thread as the internal updates are happening.

This is first step in direction of making display thread work to a
frame rate instead of fast-as-you-can. The idea will be to reduce flicker
during typing (though since the changes as part of async, this only
happens on the current line, not the whole buffer)